styletalk/requirements.txt
numpy
scikit-image
scipy
pillow
matplotlib
opencv-python
kornia
trimesh
dominate
ipython
pyyaml
ffmpeg-python
openai-whisper
pronouncing
g2p_en
tensorboard
tensorflow
torch
torchvision
face_alignment
yacs
scikit-learn
ninja
imageio
av
bournemouth-forced-aligner
